Moved sessionService to sub-directory.
