+++ src/app/cart/cart.component.js
@@ -8,7 +8,7 @@ import appConfig from 'common/app.config';
 import cartService from 'common/services/api/cart.service';
 import designationsService from 'common/services/api/designations.service';
 import modalController from 'app/productConfig/productConfig.modal';
-import sessionService from 'common/services/session.service';
+import sessionService from 'common/services/session/session.service';
 
 import template from './cart.tpl';
 import templateModal from 'app/productConfig/productConfigModal.tpl';
