+++ src/common/components/signInForm/signInForm.component.js
@@ -2,7 +2,7 @@ import angular from 'angular';
 import 'angular-gettext';
 import includes from 'lodash/includes';
 
-import sessionService from 'common/services/session.service';
+import sessionService from 'common/services/session/session.service';
 
 import template from './signInForm.tpl';
